% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combatExampleData}
\alias{combatExampleData}
\title{Example dataset for neuroCombat}
\format{
Numeric matrix with 200 rows (imaging features) 
    and 10 columns (participants).
}
\usage{
combatExampleData
}
\description{
Randomly generated measurements for 200 imaging features across
    10 participants.
}
\keyword{datasets}
