% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combatExamplePheno}
\alias{combatExamplePheno}
\title{Disease group information for combatExampleData}
\format{
Character vector of length 10 with 
    "Control" and "Disease" as possible entries.
}
\usage{
combatExamplePheno
}
\description{
Disease group information ("Control" or "Disease") for
    the 10 participants available in "combatExampleData".
}
\keyword{datasets}
