% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combatExampleScanner}
\alias{combatExampleScanner}
\title{Scanner information for combatExampleData}
\format{
Character vector of length 10 with 
    "Scanner1" and "Scanner2" as possible entries.
}
\usage{
combatExampleScanner
}
\description{
Scanner information ("Scanner 1" or "Scanner2") for
    the 10 participants available in "combatExampleData".
}
\keyword{datasets}
