% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroCombat.R
\name{neuroCombat}
\alias{neuroCombat}
\title{neuroCombat harmonization}
\usage{
neuroCombat(
  dat,
  batch,
  mod = NULL,
  eb = TRUE,
  parametric = TRUE,
  mean.only = FALSE,
  ref.batch = NULL,
  verbose = TRUE,
  BPPARAM = bpparam("SerialParam")
)
}
\arguments{
\item{dat}{Numeric matrix with imaging features as rows,
and scans/images as columns.}

\item{batch}{Numeric or character vector specifying the batch/scanner 
variable needed for harmonization. Length must be equal to the 
number of columns in \code{dat}.}

\item{mod}{Optional model matrix for outcome of interest and 
other covariates besides batch/scanner.}

\item{eb}{Should Empirical Bayes adjustements be performed? TRUE by default.
Specifying \code{eb=FALSE} will perform location/scale (L/S) adjustments
for each feature separately, that is without pooling information aross 
features to estimate batch correction parameters.}

\item{parametric}{Should parametric priors be used in the EB estimation?
TRUE by default. Note that the non-parametric option (\code{parametric=FALSE})
is usually very slow.}

\item{mean.only}{Should only correction factors be calculated for location?
FALSE by default.}

\item{ref.batch}{NULL by default.}

\item{verbose}{Should progress messages be printed? TRUE by default.}

\item{BPPARAM}{BiocParallelParam for parallel operation. 
This is mostly useful when \code{parametric=FALSE}.}
}
\value{
A named \code{list} of length 5. The 1st element (\code{dat.combat})
    contains the harmonized data. The 2nd element (\code{estimates}) contains
    estimates and other parameters used during harmonization. The 3rd element
    (\code{dataDict}) contains information about the dataset and the batch 
    covariate information that were provided to the neuroCombat function.
    The 4th element (\code{data.original}) contains the raw data that were
    provided to neuroCombat. The 5th element (\code{data.standardized})
    contains the standardized original data: each feature is scaled and
    centered after adjusting for biological covariates).
}
\description{
Main function to perform neuroCombat harmonization.
}
