% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroCombatFromTraining.R
\name{neuroCombatFromTraining}
\alias{neuroCombatFromTraining}
\title{neuroCombat harmonization using pre-trained estimates}
\usage{
neuroCombatFromTraining(dat, batch, estimates, mod = NULL, verbose = TRUE)
}
\arguments{
\item{dat}{Numeric matrix with imaging features as rows,
and scans/images as columns.}

\item{batch}{Numeric or character vector specifying the batch/scanner 
variable needed for harmonization. Length must be equal to the 
number of columns in \code{dat}.}

\item{estimates}{Prior neuroCombat estimates to be applied onto
the new dataset. Usually obtained from the output of
\code{neuroCombat} run on a training dataset.}

\item{mod}{Optional model matrix for outcome of interest and 
other covariates besides batch/scanner.}

\item{verbose}{Should progress messages be printed? TRUE by default.}
}
\value{
A named \code{list} of length 2. The first element (\code{dat.combat})
    contains the harmonized data. The second element (\code{estimates}) contains
    estimates and other parameters used during harmonization.
}
\description{
[UNDER DEVELOPMENT] neuroCombat harmonization using
    parameters estimated from a training dataset. Estimated parameters
    are applied to a new dataset, assuming scanners/batches in the 
    new dataset were also present in the training dataset.
}
