% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{standardizeData}
\alias{standardizeData}
\title{Standardize data (feature standardization)}
\usage{
standardizeData(dat, batch, mod, mean.only = FALSE)
}
\arguments{
\item{dat}{Numeric matrix with imaging features as rows,
and scans/images as columns.}

\item{batch}{Numeric or character vector specifying the batch/scanner 
variable needed for harmonization. Length must be equal to the 
number of columns in \code{dat}.}

\item{mod}{Optional model matrix for outcome of interest and 
other covariates besides batch/scanner.}

\item{mean.only}{Should only correction factors be calculated for location?
FALSE by default.}
}
\value{
Standardized data.
}
\description{
Standardize data (feature standardization). Optional
    covariates adjustment is provided.
}
